Split match and lineup rendering further
