+++ internal/ui/view_match.go
@@ -1,13 +1,10 @@
 package ui
 
 import (
-	"fmt"
 	"strings"
-	"time"
 
 	"github.com/adrunkhuman/90minuTUI/internal/site"
 	"github.com/charmbracelet/lipgloss"
-	"github.com/charmbracelet/x/ansi"
 )
 
 func (m Model) matchSketchView() string {
@@ -320,450 +317,4 @@ func (m Model) matchDetailContent(width int) string {
 	}
 
 	return strings.TrimRight(b.String(), "\n")
-}
-
-func renderBlankPanelLine(width int) string {
-	return renderFullLine("", width, colorBgPanel, colorTextSecondary, false)
-}
-
-func renderPanelRule(width int) string {
-	return renderFullLine(strings.Repeat("─", max(0, width)), width, colorBgPanel, colorBorder, false)
-}
-
-func renderCenteredPanelLine(text string, width int, fg lipgloss.Color, bold bool) string {
-	return renderFullLine(padCenter(truncate(text, width), width), width, colorBgPanel, fg, bold)
-}
-
-func matchTitleLine(home, score, away string, width int) string {
-	if score == "vs" {
-		return axisPlainLine(home, " vs ", away, width)
-	}
-	homeScore, awayScore, ok := splitScoreAxis(score)
-	if !ok || width < 32 {
-		return axisPlainLine(home+" "+score, "", away, width)
-	}
-	return axisPlainLine(home+" "+homeScore, " – ", awayScore+" "+away, width)
-}
-
-func renderScoreAxisLine(text string, width int, fg lipgloss.Color, bold bool) string {
-	left, right, ok := splitScoreAxis(text)
-	if !ok {
-		return renderCenteredPanelLine(text, width, fg, bold)
-	}
-	return renderFullLine(axisPlainLine(left, " – ", right, width), width, colorBgPanel, fg, bold)
-}
-
-func splitScoreAxis(text string) (string, string, bool) {
-	left, right, ok := strings.Cut(text, " – ")
-	if ok {
-		return strings.TrimSpace(left), strings.TrimSpace(right), true
-	}
-	left, right, ok = strings.Cut(text, "-")
-	if !ok {
-		return "", "", false
-	}
-	return strings.TrimSpace(left), strings.TrimSpace(right), true
-}
-
-func axisPlainLine(left, center, right string, width int) string {
-	centerWidth := ansi.StringWidth(center)
-	if centerWidth == 0 {
-		centerWidth = 1
-	}
-	leftWidth := max(0, (width-centerWidth)/2)
-	rightWidth := max(0, width-leftWidth-centerWidth)
-	return padLeft(truncate(left, leftWidth), leftWidth) + center + padRight(truncate(right, rightWidth), rightWidth)
-}
-
-func renderMatchMetaPanelLine(meta, weather string, width int) string {
-	parts := matchMetaDisplayParts(meta, weather)
-	if len(parts) == 0 {
-		return ""
-	}
-	return renderFullLine(matchMetaAxisLine(parts, width), width, colorBgPanel, colorTextMuted, false)
-}
-
-func matchMetaDisplayParts(meta, weather string) []string {
-	parts := matchMetaParts(meta, weather)
-	displayParts := make([]string, 0, len(parts))
-	for _, part := range parts {
-		if display := compactMatchMetaPart(part); display != "" {
-			displayParts = append(displayParts, display)
-		}
-	}
-	return displayParts
-}
-
-func matchMetaAxisLine(parts []string, width int) string {
-	if len(parts) == 0 {
-		return ""
-	}
-	parts = fitMatchMetaParts(parts, width)
-	if len(parts) == 1 {
-		return padCenter(truncate(parts[0], width), width)
-	}
-
-	separator := "  ·  "
-	if len(parts)%2 == 0 {
-		half := len(parts) / 2
-		return axisPlainLine(strings.Join(parts[:half], separator), separator, strings.Join(parts[half:], separator), width)
-	}
-
-	middle := len(parts) / 2
-	left := strings.Join(parts[:middle], separator) + separator
-	right := separator + strings.Join(parts[middle+1:], separator)
-	return axisPlainLine(left, parts[middle], right, width)
-}
-
-func fitMatchMetaParts(parts []string, width int) []string {
-	if !matchMetaAxisTruncatesAttendance(parts, width) {
-		return parts
-	}
-
-	withoutDate := make([]string, 0, len(parts))
-	for _, part := range parts {
-		if isMatchMetaDate(part) {
-			continue
-		}
-		withoutDate = append(withoutDate, part)
-	}
-	if len(withoutDate) > 0 {
-		return withoutDate
-	}
-	return parts
-}
-
-func matchMetaAxisTruncatesAttendance(parts []string, width int) bool {
-	for _, part := range parts {
-		if !strings.HasPrefix(part, "Att. ") {
-			continue
-		}
-		return !strings.Contains(matchMetaAxisLineRaw(parts, width), part)
-	}
-	return false
-}
-
-func matchMetaAxisLineRaw(parts []string, width int) string {
-	if len(parts) == 0 {
-		return ""
-	}
-	if len(parts) == 1 {
-		return padCenter(truncate(parts[0], width), width)
-	}
-
-	separator := "  ·  "
-	if len(parts)%2 == 0 {
-		half := len(parts) / 2
-		return axisPlainLine(strings.Join(parts[:half], separator), separator, strings.Join(parts[half:], separator), width)
-	}
-
-	middle := len(parts) / 2
-	left := strings.Join(parts[:middle], separator) + separator
-	right := separator + strings.Join(parts[middle+1:], separator)
-	return axisPlainLine(left, parts[middle], right, width)
-}
-
-func isMatchMetaDate(value string) bool {
-	parsed, err := time.Parse("Mon 2 January 2006, 15:04", value)
-	return err == nil && !parsed.IsZero()
-}
-
-func compactMatchMetaPart(part string) string {
-	switch {
-	case strings.HasPrefix(part, "Attendance "):
-		return "Att. " + groupDigits(strings.TrimPrefix(part, "Attendance "))
-	case strings.HasPrefix(part, "Weather "):
-		return compactWeather(strings.TrimPrefix(part, "Weather "))
-	default:
-		return formatMatchDateWithDay(part)
-	}
-}
-
-func groupDigits(value string) string {
-	digits := strings.ReplaceAll(strings.TrimSpace(value), " ", "")
-	if digits == "" {
-		return strings.TrimSpace(value)
-	}
-	parts := make([]string, 0, (len(digits)+2)/3)
-	for len(digits) > 3 {
-		parts = append([]string{digits[len(digits)-3:]}, parts...)
-		digits = digits[:len(digits)-3]
-	}
-	parts = append([]string{digits}, parts...)
-	return strings.Join(parts, " ")
-}
-
-func compactWeather(value string) string {
-	cleaned := strings.TrimSpace(value)
-	cleaned = strings.TrimSuffix(cleaned, " C")
-	cleaned = strings.TrimSuffix(cleaned, "C")
-	if cleaned == "" || strings.Contains(cleaned, "°") {
-		return cleaned
-	}
-	return cleaned + "°"
-}
-
-func formatMatchDateWithDay(value string) string {
-	parsed, err := time.Parse("2 January 2006, 15:04", value)
-	if err != nil {
-		return value
-	}
-	return parsed.Format("Mon 2 January 2006, 15:04")
-}
-
-func halftimeScoreAlways(events []site.MatchEvent) string {
-	homeGoals := 0
-	awayGoals := 0
-	for _, event := range sortedEvents(events) {
-		if !event.HasMinute || event.Minute*100+event.Stoppage > 4599 || event.Kind != "GOAL" {
-			continue
-		}
-		if event.TeamSide == "home" {
-			homeGoals++
-		} else if event.TeamSide == "away" {
-			awayGoals++
-		}
-	}
-	return fmt.Sprintf("HT %d – %d", homeGoals, awayGoals)
-}
-
-func halftimeScoreDisplay(page *site.MatchPage) string {
-	if page == nil || len(page.Events) == 0 {
-		return "HT —"
-	}
-	return halftimeScoreAlways(page.Events)
-}
-
-type timelineRow struct {
-	home   string
-	away   string
-	marker string
-	color  lipgloss.Color
-}
-
-func splitTimelineRows(events []site.MatchEvent) ([]timelineRow, []timelineRow) {
-	firstHalf := make([]timelineRow, 0, 4)
-	secondHalf := make([]timelineRow, 0, 4)
-	for _, event := range sortedEvents(events) {
-		marker, color, ok := timelineMarker(event.Kind)
-		if !ok {
-			continue
-		}
-		label := timelineEventLabel(event)
-		if label == "" {
-			continue
-		}
-		row := timelineRow{home: "—", away: "—", marker: marker, color: color}
-		switch event.TeamSide {
-		case "home":
-			row.home = label
-		case "away":
-			row.away = label
-		default:
-			continue
-		}
-
-		if event.HasMinute && event.Minute*100+event.Stoppage <= 4599 {
-			firstHalf = append(firstHalf, row)
-		} else {
-			secondHalf = append(secondHalf, row)
-		}
-	}
-
-	return firstHalf, secondHalf
-}
-
-func timelineMarker(kind string) (string, lipgloss.Color, bool) {
-	switch kind {
-	case "GOAL":
-		return "⚽", colorAccent, true
-	case "MISS":
-		return "❌", colorLoss, true
-	case "RC":
-		return "■", colorRed, true
-	default:
-		return "", colorTextMuted, false
-	}
-}
-
-func timelineEventLabel(event site.MatchEvent) string {
-	name := ansi.Strip(trimEventMinute(event))
-	minute := formatMatchMinute(event.MinuteText)
-	if name == "" || minute == "" {
-		return ""
-	}
-	if event.TeamSide == "away" {
-		return strings.TrimSpace(minute) + " " + name
-	}
-	return name + " " + minute
-}
-
-func renderTimelineRow(row timelineRow, width int) string {
-	if row.home == "—" && row.away == "—" {
-		return renderCenteredPanelLine("—", width, colorTextMuted, false)
-	}
-	left := ""
-	right := ""
-	if row.home != "—" {
-		left = row.home
-	}
-	if row.away != "—" {
-		right = row.away
-	}
-	return renderFullLine(axisPlainLine(left, " "+row.marker+" ", right, width), width, colorBgPanel, row.color, false)
-}
-
-func renderLineupsLabel(width int) string {
-	return renderSectionLabel("LINEUPS", width)
-}
-
-func renderSectionLabel(label string, width int) string {
-	label = strings.ToUpper(label)
-	line := []rune(strings.Repeat("─", max(0, width)))
-	if len(line) == 0 {
-		return ""
-	}
-	center := width / 2
-	labelStart := clamp(center-3, 0, max(0, width-len([]rune(label))))
-	for i, r := range label {
-		if labelStart+i >= len(line) {
-			break
-		}
-		line[labelStart+i] = r
-	}
-	return renderFullLine(string(line), width, colorBgPanel, colorTextMuted, false)
-}
-
-func renderLineupTwoColumnRow(home, away string, width int, bg lipgloss.Color, homeColor lipgloss.Color, awayColor lipgloss.Color, bold bool) string {
-	fg := homeColor
-	if homeColor != awayColor {
-		fg = colorTextSecondary
-	}
-	return renderFullLine(axisPlainLine(home, "│", away, width), width, bg, fg, bold)
-}
-
-func renderLineupHeaderRow(home, away string, width int) string {
-	return renderFullLine(axisPlainLine(home, "   ", away, width), width, colorBgHeader, colorAccent, true)
-}
-
-func lineupDisplayName(entry lineupEntry, side string, maxWidth int) string {
-	return formatLineupPlayerWithCards(entry, side, maxWidth, true)
-}
-
-func lineupCardFor(player site.PlayerLine, idx map[string][]site.MatchEvent) lineupCardMarker {
-	return cardMarkerAnnotationName(player.Name, idx)
-}
-
-func cardMarkerAnnotationName(name string, idx map[string][]site.MatchEvent) lineupCardMarker {
-	return cardMarkerFromEvents(matchingPlayerEvents(name, idx))
-}
-
-func substituteCardMarkerAnnotationName(name string, idx map[string][]site.MatchEvent) lineupCardMarker {
-	return cardMarkerFromEvents(matchingSubstituteCardEvents(name, idx))
-}
-
-func substituteCardMarkerAnnotationNameInRoster(name string, idx map[string][]site.MatchEvent, players []site.PlayerLine) lineupCardMarker {
-	return cardMarkerFromEvents(matchingSubstituteCardEventsInRoster(name, idx, players))
-}
-
-func cardMarkerFromEvents(matched []site.MatchEvent) lineupCardMarker {
-	if len(matched) == 0 {
-		return lineupCardMarker{}
-	}
-
-	hasYellow := false
-	for _, event := range matched {
-		switch event.Kind {
-		case "RC":
-			return lineupCardMarker{color: colorRed, ok: true}
-		case "YC":
-			hasYellow = true
-		}
-	}
-	if hasYellow {
-		return lineupCardMarker{color: colorYellow, ok: true}
-	}
-	return lineupCardMarker{}
-}
-
-func lineupCardForEntry(entry lineupEntry, idx map[string][]site.MatchEvent) lineupCardMarker {
-	return lineupCardFor(entry.player, idx)
-}
-
-func renderLineupPlayerRow(home string, homeCard lineupCardMarker, away string, awayCard lineupCardMarker, width int) string {
-	leftWidth := max(1, width/2)
-	rightWidth := max(1, width-leftWidth-1)
-	defaultStyle := lipgloss.NewStyle().Foreground(colorTextSecondary).Background(colorBgPanel)
-	divider := defaultStyle.Render("│")
-
-	left := lineupSideCell(home, homeCard, leftWidth, "home", defaultStyle)
-	right := lineupSideCell(away, awayCard, rightWidth, "away", defaultStyle)
-	return left + divider + right
-}
-
-func lineupPlayerNameWidth(width int) int {
-	leftWidth := max(1, width/2)
-	rightWidth := max(1, width-leftWidth-1)
-	return max(1, min(leftWidth, rightWidth)-2)
-}
-
-func lineupSideCell(name string, card lineupCardMarker, width int, side string, defaultStyle lipgloss.Style) string {
-	if width <= 0 {
-		return ""
-	}
-	cardWidth := 2
-	nameWidth := max(0, width-cardWidth)
-	cardStyle := lipgloss.NewStyle().Foreground(card.color).Background(colorBgPanel).Bold(true)
-
-	switch side {
-	case "home":
-		text := renderLineupLabel(name, nameWidth, "home", defaultStyle)
-		if !card.ok {
-			return text + defaultStyle.Render("  ")
-		}
-		return text + cardStyle.Render("■") + defaultStyle.Render(" ")
-	default:
-		text := renderLineupLabel(name, nameWidth, "away", defaultStyle)
-		if !card.ok {
-			return defaultStyle.Render("  ") + text
-		}
-		return defaultStyle.Render(" ") + cardStyle.Render("■") + text
-	}
-}
-
-func renderLineupLabel(label string, width int, side string, defaultStyle lipgloss.Style) string {
-	text := truncate(label, width)
-	if side == "home" {
-		text = padLeft(text, width)
-	} else {
-		text = padRight(text, width)
-	}
-
-	noteStyle := lipgloss.NewStyle().Foreground(colorTextMuted).Background(colorBgPanel)
-	yellowCardStyle := lipgloss.NewStyle().Foreground(colorYellow).Background(colorBgPanel).Bold(true)
-	redCardStyle := lipgloss.NewStyle().Foreground(colorRed).Background(colorBgPanel).Bold(true)
-	var b strings.Builder
-	inNote := false
-	for _, r := range text {
-		switch string(r) {
-		case lineupYellowCardToken:
-			b.WriteString(yellowCardStyle.Render("■"))
-			continue
-		case lineupRedCardToken:
-			b.WriteString(redCardStyle.Render("■"))
-			continue
-		}
-		if r == '(' {
-			inNote = true
-		}
-		style := defaultStyle
-		if inNote {
-			style = noteStyle
-		}
-		b.WriteString(style.Render(string(r)))
-		if r == ')' {
-			inNote = false
-		}
-	}
-	return b.String()
 }
